requirements.txt
beautifulsoup4
selenium
gspread
oauth2client
python-dotenv
